requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.0
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pip==21.1.2
pygame==2.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
setuptools==57.0.0
six==1.16.0
wheel==0.36.2
zipp==3.15.0
